backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
gtts==2.4.0
google-generativeai==0.3.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

langdetect==1.0.9
nltk==3.8.1
textblob==0.17.1